libs/sync: Move into internal
