+++ internal/cmap/cmap.go
@@ -1,7 +1,7 @@
 package cmap
 
 import (
-	cmtsync "github.com/cometbft/cometbft/libs/sync"
+	cmtsync "github.com/cometbft/cometbft/internal/sync"
 )
 
 // CMap is a goroutine-safe map
