+++ internal/events/events.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 
 	"github.com/cometbft/cometbft/internal/service"
-	cmtsync "github.com/cometbft/cometbft/libs/sync"
+	cmtsync "github.com/cometbft/cometbft/internal/sync"
 )
 
 // ErrListenerWasRemoved is returned by AddEvent if the listener was removed.
