+++ internal/flowrate/flowrate.go
@@ -10,7 +10,7 @@ import (
 	"math"
 	"time"
 
-	cmtsync "github.com/cometbft/cometbft/libs/sync"
+	cmtsync "github.com/cometbft/cometbft/internal/sync"
 )
 
 // Monitor monitors and limits the transfer rate of a data stream.
