+++ internal/statesync/syncer.go
@@ -10,8 +10,8 @@ import (
 	abci "github.com/cometbft/cometbft/abci/types"
 	"github.com/cometbft/cometbft/config"
 	sm "github.com/cometbft/cometbft/internal/state"
+	cmtsync "github.com/cometbft/cometbft/internal/sync"
 	"github.com/cometbft/cometbft/libs/log"
-	cmtsync "github.com/cometbft/cometbft/libs/sync"
 	"github.com/cometbft/cometbft/light"
 	"github.com/cometbft/cometbft/p2p"
 	ssproto "github.com/cometbft/cometbft/proto/tendermint/statesync"
