+++ internal/statesync/syncer_test.go
@@ -13,8 +13,8 @@ import (
 	"github.com/cometbft/cometbft/config"
 	sm "github.com/cometbft/cometbft/internal/state"
 	"github.com/cometbft/cometbft/internal/statesync/mocks"
+	cmtsync "github.com/cometbft/cometbft/internal/sync"
 	"github.com/cometbft/cometbft/libs/log"
-	cmtsync "github.com/cometbft/cometbft/libs/sync"
 	"github.com/cometbft/cometbft/p2p"
 	p2pmocks "github.com/cometbft/cometbft/p2p/mocks"
 	cmtstate "github.com/cometbft/cometbft/proto/tendermint/state"
