+++ privval/signer_endpoint.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/cometbft/cometbft/internal/protoio"
 	"github.com/cometbft/cometbft/internal/service"
-	cmtsync "github.com/cometbft/cometbft/libs/sync"
+	cmtsync "github.com/cometbft/cometbft/internal/sync"
 	privvalproto "github.com/cometbft/cometbft/proto/tendermint/privval"
 )
