+++ rpc/jsonrpc/client/http_json_client.go
@@ -11,7 +11,7 @@ import (
 	"net/url"
 	"strings"
 
-	cmtsync "github.com/cometbft/cometbft/libs/sync"
+	cmtsync "github.com/cometbft/cometbft/internal/sync"
 	types "github.com/cometbft/cometbft/rpc/jsonrpc/types"
 )
 
@@ -136,8 +136,10 @@ var _ HTTPClient = (*Client)(nil)
 
 // Both Client and RequestBatch can facilitate calls to the JSON
 // RPC endpoint.
-var _ Caller = (*Client)(nil)
-var _ Caller = (*RequestBatch)(nil)
+var (
+	_ Caller = (*Client)(nil)
+	_ Caller = (*RequestBatch)(nil)
+)
 
 var _ fmt.Stringer = (*Client)(nil)
