+++ rpc/jsonrpc/client/ws_client.go
@@ -14,8 +14,8 @@ import (
 
 	cmtrand "github.com/cometbft/cometbft/internal/rand"
 	"github.com/cometbft/cometbft/internal/service"
+	cmtsync "github.com/cometbft/cometbft/internal/sync"
 	"github.com/cometbft/cometbft/libs/log"
-	cmtsync "github.com/cometbft/cometbft/libs/sync"
 	types "github.com/cometbft/cometbft/rpc/jsonrpc/types"
 )
